refactor(provider): extract shared JSON flattener
Move TryParseJson/ExtractValues into an internal JsonFlattener so both
the Secrets Manager provider and the upcoming SSM provider share the
same JSON flattening rules. No behavior change.
