+++ src/AWSSecretsManager.Provider/Internal/SecretsManagerConfigurationProvider.cs
@@ -141,7 +141,7 @@ public class SecretsManagerConfigurationProvider : ConfigurationProvider, IDispo
                 {
                     _loadedValues = newValues;
                     SetData(_loadedValues, triggerReload: true);
-                    
+
                     var addedCount = newValues.Except(oldValues).Count();
                     var removedCount = oldValues.Except(newValues).Count();
                     _logger.Information("Secret changes detected and reloaded. {AddedCount} added, {RemovedCount} removed",
@@ -167,99 +167,6 @@ public class SecretsManagerConfigurationProvider : ConfigurationProvider, IDispo
             {
                 _loadedValues = newValues;
                 SetData(_loadedValues, triggerReload: true);
-            }
-        }
-    }
-
-    private static bool TryParseJson(string data, out JsonElement? jsonElement)
-    {
-        jsonElement = null;
-
-        data = data.TrimStart();
-        var firstChar = data.FirstOrDefault();
-
-        if (firstChar != '[' && firstChar != '{')
-        {
-            return false;
-        }
-
-        try
-        {
-            using var jsonDocument = JsonDocument.Parse(data);
-            //  https://docs.microsoft.com/en-us/dotnet/standard/serialization/system-text-json-use-dom-utf8jsonreader-utf8jsonwriter?pivots=dotnet-6-0#jsondocument-is-idisposable
-            //  Its recommended to return the clone of the root element as the json document will be disposed
-            jsonElement = jsonDocument.RootElement.Clone();
-            return true;
-        }
-        catch (JsonException)
-        {
-            return false;
-        }
-    }
-
-    private static IEnumerable<(string key, string? value)> ExtractValues(JsonElement? jsonElement, string prefix)
-    {
-        if (jsonElement == null)
-        {
-            yield break;
-        }
-        var element = jsonElement.Value;
-        switch (element.ValueKind)
-        {
-            case JsonValueKind.Array:
-            {
-                var currentIndex = 0;
-                foreach (var el in element.EnumerateArray())
-                {
-                    var secretKey = $"{prefix}{ConfigurationPath.KeyDelimiter}{currentIndex}";
-                    foreach (var (key, value) in ExtractValues(el, secretKey))
-                    {
-                        yield return (key, value);
-                    }
-                    currentIndex++;
-                }
-                break;
-            }
-            case JsonValueKind.Number:
-            {
-                var value = element.GetRawText();
-                yield return (prefix, value);
-                break;
-            }
-            case JsonValueKind.String:
-            {
-                var value = element.GetString() ?? "";
-                yield return (prefix, value);
-                break;
-            }
-            case JsonValueKind.True:
-            case JsonValueKind.False:
-            {
-                var value = element.GetBoolean();
-                yield return (prefix, value.ToString());
-                break;
-            }
-            case JsonValueKind.Object:
-            {
-                foreach (var property in element.EnumerateObject())
-                {
-                    var secretKey = $"{prefix}{ConfigurationPath.KeyDelimiter}{property.Name}";
-                    foreach (var (key, value) in ExtractValues(property.Value, secretKey))
-                    {
-                        yield return (key, value);
-                    }
-                }
-                break;
-            }
-            case JsonValueKind.Null:
-            {
-                yield return (prefix, null);
-                break;
-            }
-            case JsonValueKind.Undefined:
-            default:
-            {
-                throw new FormatException("unsupported json token");
             }
         }
     }
@@ -335,10 +242,10 @@ public class SecretsManagerConfigurationProvider : ConfigurationProvider, IDispo
                 if (secretString is null)
                     continue;
 
-                if (TryParseJson(secretString, out var jElement))
+                if (JsonFlattener.TryParseJson(secretString, out var jElement))
                 {
                     // [MaybeNullWhen(false)] attribute is available in .net standard since version 2.1
-                    var values = ExtractValues(jElement!, secretName);
+                    var values = JsonFlattener.ExtractValues(jElement!, secretName);
 
                     foreach (var (key, value) in values)
                     {
@@ -434,10 +341,10 @@ public class SecretsManagerConfigurationProvider : ConfigurationProvider, IDispo
                     if (secretString is null)
                         continue;
 
-                    if (TryParseJson(secretString, out var jElement))
+                    if (JsonFlattener.TryParseJson(secretString, out var jElement))
                     {
                         // [MaybeNullWhen(false)] attribute is available in .net standard since version 2.1
-                        var values = ExtractValues(jElement!, secretName);
+                        var values = JsonFlattener.ExtractValues(jElement!, secretName);
 
                         foreach (var (key, value) in values)
                         {
